[ASSIGNOR] refactor Assignor
