+++ common/src/main/java/org/astraea/common/assignor/Assignor.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.common.consumer.assignor;
+package org.astraea.common.assignor;
 
 import java.net.InetSocketAddress;
 import java.time.Duration;
@@ -22,13 +22,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import org.apache.kafka.common.Cluster;
+import org.apache.kafka.common.Configurable;
 import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
-import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.NodeInfo;
-import org.astraea.common.admin.ReplicaInfo;
 import org.astraea.common.admin.TopicPartition;
 import org.astraea.common.cost.HasPartitionCost;
 import org.astraea.common.cost.ReplicaSizeCost;
@@ -36,7 +37,8 @@ import org.astraea.common.metrics.collector.MetricCollector;
 import org.astraea.common.partitioner.PartitionerUtils;
 
 /** Abstract assignor implementation which does some common work (e.g., configuration). */
-public abstract class AbstractConsumerPartitionAssignor implements ConsumerPartitionAssignor {
+public abstract class Assignor
+    implements org.apache.kafka.clients.consumer.ConsumerPartitionAssignor, Configurable {
   public static final String JMX_PORT = "jmx.port";
   Function<Integer, Optional<Integer>> jmxPortGetter = (id) -> Optional.empty();
   HasPartitionCost costFunction = HasPartitionCost.EMPTY;
@@ -50,34 +52,26 @@ public abstract class AbstractConsumerPartitionAssignor implements ConsumerParti
           .build();
 
   /**
-   * Perform the group assignment given the members' subscription and ClusterInfo.
+   * Perform the group assignment given the member subscriptions and current cluster metadata.
    *
    * @param subscriptions Map from the member id to their respective topic subscription.
-   * @param metadata Current topic/broker metadata known by consumer.
-   * @return Map from each member to the list of topic-partitions assigned to them.
+   * @param topicPartitions Current topic/broker metadata known by consumer.
+   * @return Map from each member to the list of partitions assigned to them.
    */
-  public abstract Map<String, List<TopicPartition>> assign(
-      Map<String, org.astraea.common.consumer.assignor.Subscription> subscriptions,
-      ClusterInfo<ReplicaInfo> metadata);
+  protected abstract Map<String, List<TopicPartition>> assign(
+      Map<String, org.astraea.common.assignor.Subscription> subscriptions,
+      Set<TopicPartition> topicPartitions);
+  // TODO: replace the topicPartitions by ClusterInfo after Assignor is able to handle Admin
+  // https://github.com/skiptests/astraea/issues/1409
 
   /**
    * Parse config to get JMX port and cost function type.
    *
    * @param config configuration
    */
-  @Override
-  public void configure(Configuration config) {
-    var costFunctions = parseCostFunctionWeight(config);
-    var customJMXPort = PartitionerUtils.parseIdJMXPort(config);
-    var defaultJMXPort = config.integer(JMX_PORT);
+  protected void configure(Configuration config) {}
 
-    this.costFunction =
-        costFunctions.isEmpty()
-            ? HasPartitionCost.of(Map.of(new ReplicaSizeCost(), 1D))
-            : HasPartitionCost.of(costFunctions);
-    this.jmxPortGetter = id -> Optional.ofNullable(customJMXPort.get(id)).or(() -> defaultJMXPort);
-    this.costFunction.fetcher().ifPresent(metricCollector::addFetcher);
-  }
+  // -----------------------[helper]-----------------------//
 
   /**
    * check the nodes which wasn't register yet.
@@ -138,5 +132,51 @@ public abstract class AbstractConsumerPartitionAssignor implements ConsumerParti
             Collectors.toMap(
                 e -> Utils.construct((Class<HasPartitionCost>) e.getKey(), config),
                 Map.Entry::getValue));
+  }
+
+  // -----------------------[kafka method]-----------------------//
+
+  @Override
+  public final GroupAssignment assign(Cluster metadata, GroupSubscription groupSubscription) {
+    // convert Kafka's data structure to ours
+    var subscriptionsPerMember =
+        org.astraea.common.assignor.GroupSubscription.from(groupSubscription).groupSubscription();
+
+    var topicPartitions =
+        metadata.topics().stream()
+            .flatMap(
+                name ->
+                    metadata.partitionsForTopic(name).stream()
+                        .map(p -> TopicPartition.of(p.topic(), p.partition())))
+            .collect(Collectors.toSet());
+
+    return new GroupAssignment(
+        assign(subscriptionsPerMember, topicPartitions).entrySet().stream()
+            .collect(
+                Collectors.toMap(
+                    Map.Entry::getKey,
+                    e ->
+                        new Assignment(
+                            e.getValue().stream()
+                                .map(TopicPartition::to)
+                                .collect(Collectors.toUnmodifiableList())))));
+  }
+
+  @Override
+  public final void configure(Map<String, ?> configs) {
+    var config =
+        Configuration.of(
+            configs.entrySet().stream()
+                .collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().toString())));
+    var costFunctions = parseCostFunctionWeight(config);
+    var customJMXPort = PartitionerUtils.parseIdJMXPort(config);
+    var defaultJMXPort = config.integer(JMX_PORT);
+    this.costFunction =
+        costFunctions.isEmpty()
+            ? HasPartitionCost.of(Map.of(new ReplicaSizeCost(), 1D))
+            : HasPartitionCost.of(costFunctions);
+    this.jmxPortGetter = id -> Optional.ofNullable(customJMXPort.get(id)).or(() -> defaultJMXPort);
+    this.costFunction.fetcher().ifPresent(metricCollector::addFetcher);
+    configure(config);
   }
 }
