--- common/src/main/java/org/astraea/common/consumer/assignor/RandomAssignor.java
@@ -1,53 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.astraea.common.consumer.assignor;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import org.astraea.common.admin.ClusterInfo;
-import org.astraea.common.admin.ReplicaInfo;
-import org.astraea.common.admin.TopicPartition;
-
-public class RandomAssignor extends AbstractConsumerPartitionAssignor {
-
-  @Override
-  public Map<String, List<TopicPartition>> assign(
-      Map<String, org.astraea.common.consumer.assignor.Subscription> subscriptions,
-      ClusterInfo<ReplicaInfo> metadata) {
-    Set<TopicPartition> topicPartitions = metadata.topicPartitions();
-
-    var assignments = new HashMap<String, List<TopicPartition>>();
-    var consumers = new ArrayList<>(subscriptions.keySet());
-    consumers.forEach(consumer -> assignments.put(consumer, new ArrayList<>()));
-
-    topicPartitions.forEach(
-        topicPartition -> {
-          var consumer = consumers.get((int) (Math.random() * consumers.size()));
-          assignments.get(consumer).add(topicPartition);
-        });
-
-    return assignments;
-  }
-
-  @Override
-  public String name() {
-    return "random";
-  }
-}
