+++ common/src/test/java/org/astraea/common/assignor/AssignorTest.java
@@ -14,26 +14,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.common.consumer.assignor;
+package org.astraea.common.assignor;
 
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import org.apache.kafka.common.TopicPartition;
 import org.astraea.common.Configuration;
+import org.astraea.common.admin.NodeInfo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class ConsumerPartitionAssignorTest {
+public class AssignorTest {
   @Test
   void testSubscriptionConvert() {
     var data = "rack=1";
     var userData = ByteBuffer.wrap(data.getBytes(StandardCharsets.UTF_8));
     var kafkaSubscription =
-        new ConsumerPartitionAssignor.Subscription(List.of("test"), userData, null);
-    var ourSubscription = org.astraea.common.consumer.assignor.Subscription.from(kafkaSubscription);
+        new org.apache.kafka.clients.consumer.ConsumerPartitionAssignor.Subscription(
+            List.of("test"), userData, null);
+    var ourSubscription = Subscription.from(kafkaSubscription);
 
     Assertions.assertEquals(kafkaSubscription.topics(), ourSubscription.topics());
     Assertions.assertEquals(kafkaSubscription.ownedPartitions(), ourSubscription.ownedPartitions());
@@ -45,19 +48,18 @@ public class ConsumerPartitionAssignorTest {
   @Test
   void testGroupSubscriptionConvert() {
     var kafkaUser1Subscription =
-        new ConsumerPartitionAssignor.Subscription(
+        new org.apache.kafka.clients.consumer.ConsumerPartitionAssignor.Subscription(
             List.of("test1", "test2"), convert("rack=1"), null);
     var kafkaUser2Subscription =
-        new ConsumerPartitionAssignor.Subscription(
+        new org.apache.kafka.clients.consumer.ConsumerPartitionAssignor.Subscription(
             List.of("test1", "test2"),
             convert("rack=2"),
             List.of(new TopicPartition("test1", 0), new TopicPartition("test2", 1)));
     kafkaUser2Subscription.setGroupInstanceId(Optional.of("astraea"));
     var kafkaGroupSubscription =
-        new ConsumerPartitionAssignor.GroupSubscription(
+        new org.apache.kafka.clients.consumer.ConsumerPartitionAssignor.GroupSubscription(
             Map.of("user1", kafkaUser1Subscription, "user2", kafkaUser2Subscription));
-    var ourGroupSubscription =
-        org.astraea.common.consumer.assignor.GroupSubscription.from(kafkaGroupSubscription);
+    var ourGroupSubscription = GroupSubscription.from(kafkaGroupSubscription);
 
     var ourUser1Subscription = ourGroupSubscription.groupSubscription().get("user1");
     var ourUser2Subscription = ourGroupSubscription.groupSubscription().get("user2");
@@ -99,5 +101,60 @@ public class ConsumerPartitionAssignorTest {
 
   private static ByteBuffer convert(String value) {
     return ByteBuffer.wrap(value.getBytes(StandardCharsets.UTF_8));
+  }
+
+  @Test
+  void testUnregisterId() {
+    var assignor = new RandomAssignor();
+    assignor.configure(Map.of("broker.1000.jmx.port", "8000", "broker.1001.jmx.port", "8100"));
+    var nodes =
+        List.of(NodeInfo.of(1000, "192.168.103.1", 8000), NodeInfo.of(1001, "192.168.103.2", 8100));
+    var unregister = assignor.checkUnregister(nodes);
+    Assertions.assertEquals(2, unregister.size());
+    Assertions.assertEquals("192.168.103.1", unregister.get(1000));
+    Assertions.assertEquals("192.168.103.2", unregister.get(1001));
+
+    assignor.registerLocalJMX(Map.of(1000, "192.168.103.1"));
+    var unregister2 = assignor.checkUnregister(nodes);
+    Assertions.assertEquals(1, unregister2.size());
+    Assertions.assertEquals("192.168.103.2", unregister2.get(1001));
+  }
+
+  @Test
+  void testAddNode() {
+    var assignor = new RandomAssignor();
+    assignor.configure(Map.of("broker.1000.jmx.port", "8000"));
+    var nodes = new ArrayList<NodeInfo>();
+    nodes.add(NodeInfo.of(1000, "192.168.103.1", 8000));
+    var unregisterNode = assignor.checkUnregister(nodes);
+    Assertions.assertEquals(1, unregisterNode.size());
+    Assertions.assertEquals("192.168.103.1", unregisterNode.get(1000));
+    assignor.registerLocalJMX(unregisterNode);
+    unregisterNode = assignor.checkUnregister(nodes);
+    Assertions.assertEquals(0, unregisterNode.size());
+
+    // after add a new node, assignor check unregister node
+
+    nodes.add(NodeInfo.of(1001, "192.168.103.2", 8000));
+    unregisterNode = assignor.checkUnregister(nodes);
+    Assertions.assertEquals(1, unregisterNode.size());
+    Assertions.assertEquals("192.168.103.2", unregisterNode.get(1001));
+  }
+
+  @Test
+  void testParseCostFunctionWeight() {
+    var costFunction =
+        Assignor.parseCostFunctionWeight(
+            Configuration.of(Map.of("org.astraea.common.cost.ReplicaSizeCost", "100")));
+    Assertions.assertEquals(1, costFunction.size());
+    for (var e : costFunction.entrySet()) {
+      Assertions.assertEquals(
+          "org.astraea.common.cost.ReplicaSizeCost", e.getKey().getClass().getName());
+      Assertions.assertEquals(100, e.getValue());
+    }
+
+    var negativeConfig = Configuration.of(Map.of("org.astraea.common.cost.ReplicaSizeCost", "-1"));
+    Assertions.assertThrows(
+        IllegalArgumentException.class, () -> Assignor.parseCostFunctionWeight(negativeConfig));
   }
 }
